reduced PizzaStores Factory Method example to fewer files
