+++ Design-Patterns/Factory-Pattern/PizzaStore.cs
@@ -1,5 +1,4 @@
-﻿using Factory_Pattern.Pizzas;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -21,6 +20,60 @@ namespace Factory_Pattern
             pizza.Bake();
             pizza.Cut();
             pizza.Box();
+
+            return pizza;
+        }
+    }
+
+    internal class ChicagoPizzaStore : PizzaStore
+    {
+        protected override Pizza CreatePizza(string type)
+        {
+            Pizza pizza;
+
+            if (type == "cheese")
+            {
+                pizza = new CheesePizza("Chicago");
+            }
+            else if (type == "pepperoni")
+            {
+                pizza = new PepperoniPizza("Chicago");
+            }
+            else if (type == "clam")
+            {
+                pizza = new ClamPizza("Chicago");
+            }
+            else
+            {
+                pizza = new VeggiePizza("Chicago");
+            }
+
+            return pizza;
+        }
+    }
+
+    internal class NYPizzaStore : PizzaStore
+    {
+        protected override Pizza CreatePizza(string type)
+        {
+            Pizza pizza;
+
+            if (type == "cheese")
+            {
+                pizza = new CheesePizza("New York");
+            }
+            else if (type == "pepperoni")
+            {
+                pizza = new PepperoniPizza("New York");
+            }
+            else if (type == "clam")
+            {
+                pizza = new ClamPizza("New York");
+            }
+            else
+            {
+                pizza = new VeggiePizza("New York");
+            }
 
             return pizza;
         }
