--- Design-Patterns/Factory-Pattern/PizzaStores/NYPizzaStore.cs
@@ -1,36 +1,0 @@
-﻿using Factory_Pattern.Pizzas;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Factory_Pattern.PizzaStores
-{
-    internal class NYPizzaStore : PizzaStore
-    {
-        protected override Pizza CreatePizza(string type)
-        {
-            Pizza pizza;
-
-            if (type == "cheese")
-            {
-                pizza = new CheesePizza("New York");
-            }
-            else if (type == "pepperoni")
-            {
-                pizza = new PepperoniPizza("New York");
-            }
-            else if (type == "clam")
-            {
-                pizza = new ClamPizza("New York");
-            }
-            else
-            {
-                pizza = new VeggiePizza("New York");
-            }
-
-            return pizza;
-        }
-    }
-}
