--- Design-Patterns/Factory-Pattern/Pizzas/ClamPizza.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Factory_Pattern.Pizzas
-{
-    internal class ClamPizza : Pizza
-    {
-        public ClamPizza(string style) : base(style) { }
-    }
-}
