--- Design-Patterns/Factory-Pattern/Pizzas/PepperoniPizza.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Factory_Pattern.Pizzas
-{
-    internal class PepperoniPizza : Pizza
-    {
-        public PepperoniPizza(string style) : base(style) { }
-    }
-}
