+++ Design-Patterns/Factory-Pattern/Program.cs
@@ -1,5 +1,4 @@
-﻿using Factory_Pattern.PizzaStores;
-
+﻿
 namespace Factory_Pattern
 {
     internal class Program
