--- Design-Patterns/Factory-Pattern/SimplePizzaFactory.cs
@@ -1,14 +1,0 @@
-﻿using Factory_Pattern.Pizzas;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Factory_Pattern
-{
-    abstract class SimplePizzaFactory
-    {
-        public abstract Pizza CreatePizza(string type);
-    }
-}
